requirements.txt
Flask==2.0.3
google-cloud-firestore==2.3.0
google-auth==2.6.0
